Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0
PyPDF2==3.0.1
regex==2023.10.3
openpyxl==3.1.2
pandas==2.1.3
